+++ src/sandi/attributes/PCAM.py
@@ -24,7 +24,7 @@ class PCam3_characteristics:
         """
         Class to store image real-life measurements (default values are from the PCam3 from Herbst Environmental Science).
         """
-        self.image_height = StringVar(value="14.9")
-        self.image_width = StringVar(value="22.3")
+        #self.image_height = StringVar(value="14.9")
+        #self.image_width = StringVar(value="22.3")
         self.image_depth = StringVar(value="0.405")
         self.pixel_size = StringVar(value="3.21")